requirements.txt
requests==2.31.0
python-dotenv==1.0.0
python-slugify==8.0.1
pytube==15.0.0
openai-whisper
whisperplus
